+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+[2.2]
+
+- Adds CLI function to annotate variants
+- Adds CLI functionality to dump and restore a database
+
+[2.1]
+
+- Fix bug with inserting variants
+
 [2.0]
 
 - Adds structural variants to loqus